[Bronze II] Title: 문자열 반복, Time: 0 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Bronze/2675. 문자열 반복/문자열 반복.cc
@@ -10,7 +10,7 @@ int main() {
 	cin >> T;
 	for (int i = 0; i < T; i++) {
 		cin >> R >> name;
-			for (size_t j = 0; j < name.length(); j++) {
+			for (int j = 0; j < name.length(); j++) {
 				for (int x = 0; x < R; x++) 
 					cout << name[j];
 			}
